ai_ml/requirements.txt
# AI/ML Dependencies for Training and Deployment

# Deep Learning Frameworks
torch==2.1.0
torchvision==0.16.0
tensorflow==2.14.0
onnx==1.15.0
onnxruntime==1.16.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
albumentations==1.3.1

# Data & Utils
numpy==1.24.3
pandas==2.1.0
scikit-learn==1.3.0

# Experiment Tracking
wandb==0.16.0
mlflow==2.8.0

# Model Optimization
torch-pruning==1.3.0
neural-compressor==2.4

# Dataset Tools
pytorch-lightning==2.1.0
torchmetrics==1.2.0

# API/Deployment
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Progress & Logging
tqdm==4.66.1
loguru==0.7.2

# Containerization
docker==6.1.3

# Cloud ML
google-cloud-aiplatform==1.36.0
boto3==1.28.85  # AWS SageMaker

# Model Serving
tensorflow-serving-api==2.14.0
tritonclient==2.40.0
